separate rest_server from rest_client
