+++ tests/rest_test/rest_registry_test.cpp
@@ -2,7 +2,7 @@
 // For the license information refer to glaze.hpp
 
 #include "glaze/glaze.hpp"
-#include "glaze/rest/rest.hpp"
+#include "glaze/rest/rest_server.hpp"
 #include "glaze/rpc/registry.hpp"
 #include "ut/ut.hpp"
