+++ tests/rest_test/rest_test.cpp
@@ -2,7 +2,7 @@
 // For the license information refer to glaze.hpp
 
 #include "glaze/glaze.hpp"
-#include "glaze/rest/rest.hpp"
+#include "glaze/rest/rest_server.hpp"
 #include "ut/ut.hpp"
 
 #include <iostream>
